requirements.txt
certifi==2018.4.16
chardet==3.0.4
idna==2.6
requests==2.18.4
urllib3==1.22
Flask>=0.10.1
python-telegram-bot>=2.5
gunicorn>=19.9.0
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
oauth2client
pymongo